requirements.txt
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
comm==0.1.3
cryptography==41.0.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.5
django-cors-headers==4.1.0
django-filter==23.2
django-import-export==3.2.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
executing==1.2.0
flake8==6.1.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
MarkupPy==1.14
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
numpy==1.25.0
oauth2client==4.1.3
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.8.0.74
openpyxl==3.1.2
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.7
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
stripe==5.4.0
tablib==3.5.0
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
zipp==3.16.2
